requirements.txt
requests
boto3==1.35.99
python-dotenv==0.19.2
pytz==2021.1
requests==2.32.3
s3transfer==0.10.4
urllib3==2.3.0
